// src/core/overlay/interface.ts
export type ThyPlacement =
    | 'top'
    | 'topLeft'
    | 'topRight'
    | 'bottom'
    | 'bottomLeft'
    | 'bottomRight'
    | 'left'
    | 'leftTop'
    | 'leftBottom'
    | 'right'
    | 'rightTop'
    | 'rightBottom';
export interface ConnectedPositionOffset {
    offsetY?: number;
    offsetX?: number;
}